5.0.0.16
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/branches/pre_5_0_beta@542 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/itemdbtool.cpp
@@ -1683,7 +1683,7 @@ int displayRecord(EQItemDB* itemDB,
     printf("001: ");
     uint8_t fieldCount = 1;
     uint8_t col = 5;
-    size_t fieldWidth;
+    int fieldWidth;
     const char* curPos = rawData;
     const char* endField;
